Ditch ShareLink per https://github.com/DoSomething/phoenix-next/pull/2114#discussion_r422371975
